requirements.txt
flask==3.1.0
gunicorn==22.0.0
numpy==1.26.4
Pillow==10.4.0
opencv-python-headless==4.10.0.84
torch==2.1.0
torchvision==0.16.0
